refactor: index.js -> config.js
